pyproject.toml
[project]
name = "project-cats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.116.1",
    "greenlet>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.7",
    "python-dotenv>=1.2.1",
    "redis>=5.0.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.35.0",
]

[dependency-groups]
ipynb = [
    "jupyter>=1.1.1",
    "ipywidgets>=8.1.8",
    "ipykernel>=7.1.0",
]
migrations = [
    "alembic>=1.17.2",
]
